further refactoring
